working through storeslice branches
